Dispatch activity display into the main loop
Before this change, picowota would ignore data on the tcp connection if it is
sent immediately after the connection was opened.

The reason why that happens happens:
- Client connects
- tcp_comm_client_init is called
- the gpio led turns on (tcp_comm.c#L461)
- While communicating with the cyw43 to enable the led, network packages are
  discovered that need to be processed, incl. the initial opcode.
- the recv-callback is not set yet, so the callback invocation is skipped
- tcp_comm_client_init continues to set up callbacks, the initial data is lost.

To resolve this, the code that turns the LED on and off is dispatched into the
mainloop, so it is never called from within a networking callback.