picowota: tweak ld scripts for fix strange issue in picowota binary generation
The objcopy step that updates sections with binary data failed because of an error with sections now marked NOLOAD instead of COPY.

The error disappeared with moving section .uninitialized_data to near .ram_vector_table, which somehow triggered adding a section that reserved a piece of RAM (which is fine) with reference to a flash area (which is strange).

Note that these uninitialized sections are explicitly requested by __uninitialized_ram() directives:

pico-sdk/src/rp2_common/pico_platform/include/pico/platform.h:
    #define __uninitialized_ram(group) __attribute__((section(".uninitialized_data." #group))) group

Their use is to keep ram uninitialized on purpose, to collect entropy (for RNG) or data left over from before reset.